requirements.txt
fastapi
uvicorn
pyyaml
pydantic
psutil
pytest
websockets
requests
rich
